Add files via upload
fix for [a : b : 0.5] type prompt editing with T5 and Gemma - prompt parser handles list of conds, so torch.stack is not needed, so padding of conds is not needed. (Previously, could fail if a and b were different lengths after text encoding. Flux pads to min_length 256 anyway, hiding the issue in typical use; but Chroma does not.)
also some ruffage